Move indexing related tests to `ruby_indexer.rs`
