+++ rust/saturn/src/indexing/ruby_indexer.rs
@@ -1877,4 +1877,100 @@ mod tests {
             &DeclarationId::from("Foo::g=")
         );
     }
+
+    #[test]
+    fn index_unresolved_constant_references_inside_compact_namespace() {
+        let mut context = GraphTest::new();
+
+        context.index_uri("file:///foo.rb", {
+            r"
+            module Foo
+              class Bar::Baz
+                String
+              end
+            end
+            "
+        });
+
+        let refs = context.graph.unresolved_references();
+        assert_eq!(refs.len(), 1);
+
+        let reference = &refs[0];
+
+        match reference {
+            UnresolvedReference::Constant(unresolved) => {
+                assert_eq!(unresolved.name_id(), &NameId::from("String"));
+                assert_eq!(
+                    unresolved.nesting().as_ref().unwrap().ids_as_vec(),
+                    vec![DeclarationId::from("Foo"), DeclarationId::from("Foo::Bar::Baz")]
+                );
+                assert_eq!(unresolved.uri_id(), UriId::from("file:///foo.rb"));
+                assert_eq!(unresolved.offset(), &Offset::new(32, 38));
+            }
+        }
+    }
+
+    #[test]
+    fn index_unresolved_constant_references() {
+        let mut context = GraphTest::new();
+
+        context.index_uri("file:///foo.rb", {
+            r"
+            module Foo
+              class Bar
+                String
+              end
+            end
+            "
+        });
+
+        let refs = context.graph.unresolved_references();
+        assert_eq!(refs.len(), 1);
+
+        let reference = &refs[0];
+
+        match reference {
+            UnresolvedReference::Constant(unresolved) => {
+                assert_eq!(unresolved.name_id(), &NameId::from("String"));
+                assert_eq!(
+                    unresolved.nesting().as_ref().unwrap().ids_as_vec(),
+                    vec![DeclarationId::from("Foo"), DeclarationId::from("Foo::Bar")]
+                );
+                assert_eq!(unresolved.uri_id(), UriId::from("file:///foo.rb"));
+                assert_eq!(unresolved.offset(), &Offset::new(27, 33));
+            }
+        }
+    }
+
+    #[test]
+    fn index_unresolved_constant_path_references() {
+        let mut context = GraphTest::new();
+
+        context.index_uri("file:///foo.rb", {
+            r"
+            module Foo
+              class Bar
+                Object::String
+              end
+            end
+            "
+        });
+
+        let refs = context.graph.unresolved_references();
+        assert_eq!(refs.len(), 1);
+
+        let reference = &refs[0];
+
+        match reference {
+            UnresolvedReference::Constant(unresolved) => {
+                assert_eq!(unresolved.name_id(), &NameId::from("Object::String"));
+                assert_eq!(
+                    unresolved.nesting().as_ref().unwrap().ids_as_vec(),
+                    vec![DeclarationId::from("Foo"), DeclarationId::from("Foo::Bar")]
+                );
+                assert_eq!(unresolved.uri_id(), UriId::from("file:///foo.rb"));
+                assert_eq!(unresolved.offset(), &Offset::new(27, 41));
+            }
+        }
+    }
 }
